registration/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
crispy-bootstrap5==0.6
Deprecated==1.2.13
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-redis==5.2.0
django-select2==7.10.0
Django==4.1
packaging==21.3
Pillow==9.2.0
pip==22.2.2
pymemcache==3.5.2
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
setuptools==63.3.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
wheel==0.37.1
wrapt==1.14.1
